enable error-prone "narrow calculation" check (#11923)
This check finds bugs such as https://github.com/apache/lucene/pull/11905.

See https://errorprone.info/bugpattern/NarrowCalculation

+++ lucene/core/src/java/org/apache/lucene/index/BinaryDocValuesFieldUpdates.java
@@ -138,7 +138,7 @@ final class BinaryDocValuesFieldUpdates extends DocValuesFieldUpdates {
         + lengths.ramBytesUsed()
         + RamUsageEstimator.NUM_BYTES_OBJECT_HEADER
         + 2 * Integer.BYTES
-        + 3 * RamUsageEstimator.NUM_BYTES_OBJECT_REF
+        + 3 * (long) RamUsageEstimator.NUM_BYTES_OBJECT_REF
         + values.bytes().length;
   }
 }
