+++ lucene/test-framework/src/java/org/apache/lucene/tests/store/BaseDirectoryTestCase.java
@@ -444,7 +444,7 @@ public abstract class BaseDirectoryTestCase extends LuceneTestCase {
           longs[i] = random().nextBoolean() ? Long.MIN_VALUE : Long.MAX_VALUE;
           break;
         case 2:
-          longs[i] = (random().nextBoolean() ? -1 : 1) * random().nextInt(1024);
+          longs[i] = (random().nextBoolean() ? -1 : 1) * (long) random().nextInt(1024);
           break;
         default:
           throw new AssertionError();
@@ -685,8 +685,8 @@ public abstract class BaseDirectoryTestCase extends LuceneTestCase {
       o.close();
 
       IndexInput i = dir.openInput("out", newIOContext(random()));
-      i.seek(2 * bufferLength - 1);
-      i.seek(3 * bufferLength);
+      i.seek(2L * bufferLength - 1);
+      i.seek(3L * bufferLength);
       i.seek(bufferLength);
       i.readBytes(bytes, 0, 2 * bufferLength);
       i.close();
@@ -956,15 +956,15 @@ public abstract class BaseDirectoryTestCase extends LuceneTestCase {
       IndexInput input = dir.openInput("longs", newIOContext(random()));
       RandomAccessInput slice = input.randomAccessSlice(0, input.length());
       for (int i = 0; i < longs.length; i++) {
-        assertEquals(longs[i], slice.readLong(i * 8));
+        assertEquals(longs[i], slice.readLong(i * 8L));
       }
 
       // subslices
       for (int i = 1; i < longs.length; i++) {
-        long offset = i * 8;
+        long offset = i * 8L;
         RandomAccessInput subslice = input.randomAccessSlice(offset, input.length() - offset);
         for (int j = i; j < longs.length; j++) {
-          assertEquals(longs[j], subslice.readLong((j - i) * 8));
+          assertEquals(longs[j], subslice.readLong((j - i) * 8L));
         }
       }
 
@@ -981,7 +981,7 @@ public abstract class BaseDirectoryTestCase extends LuceneTestCase {
         IndexInput padded = dir.openInput(name, newIOContext(random()));
         RandomAccessInput whole = padded.randomAccessSlice(i, padded.length() - i);
         for (int j = 0; j < longs.length; j++) {
-          assertEquals(longs[j], whole.readLong(j * 8));
+          assertEquals(longs[j], whole.readLong(j * 8L));
         }
         padded.close();
       }
@@ -1005,15 +1005,15 @@ public abstract class BaseDirectoryTestCase extends LuceneTestCase {
       IndexInput input = dir.openInput("ints", newIOContext(random()));
       RandomAccessInput slice = input.randomAccessSlice(0, input.length());
       for (int i = 0; i < ints.length; i++) {
-        assertEquals(ints[i], slice.readInt(i * 4));
+        assertEquals(ints[i], slice.readInt(i * 4L));
       }
 
       // subslices
       for (int i = 1; i < ints.length; i++) {
-        long offset = i * 4;
+        long offset = i * 4L;
         RandomAccessInput subslice = input.randomAccessSlice(offset, input.length() - offset);
         for (int j = i; j < ints.length; j++) {
-          assertEquals(ints[j], subslice.readInt((j - i) * 4));
+          assertEquals(ints[j], subslice.readInt((j - i) * 4L));
         }
       }
 
@@ -1030,7 +1030,7 @@ public abstract class BaseDirectoryTestCase extends LuceneTestCase {
         IndexInput padded = dir.openInput(name, newIOContext(random()));
         RandomAccessInput whole = padded.randomAccessSlice(i, padded.length() - i);
         for (int j = 0; j < ints.length; j++) {
-          assertEquals(ints[j], whole.readInt(j * 4));
+          assertEquals(ints[j], whole.readInt(j * 4L));
         }
         padded.close();
       }
@@ -1053,15 +1053,15 @@ public abstract class BaseDirectoryTestCase extends LuceneTestCase {
       IndexInput input = dir.openInput("shorts", newIOContext(random()));
       RandomAccessInput slice = input.randomAccessSlice(0, input.length());
       for (int i = 0; i < shorts.length; i++) {
-        assertEquals(shorts[i], slice.readShort(i * 2));
+        assertEquals(shorts[i], slice.readShort(i * 2L));
       }
 
       // subslices
       for (int i = 1; i < shorts.length; i++) {
-        long offset = i * 2;
+        long offset = i * 2L;
         RandomAccessInput subslice = input.randomAccessSlice(offset, input.length() - offset);
         for (int j = i; j < shorts.length; j++) {
-          assertEquals(shorts[j], subslice.readShort((j - i) * 2));
+          assertEquals(shorts[j], subslice.readShort((j - i) * 2L));
         }
       }
 
@@ -1078,7 +1078,7 @@ public abstract class BaseDirectoryTestCase extends LuceneTestCase {
         IndexInput padded = dir.openInput(name, newIOContext(random()));
         RandomAccessInput whole = padded.randomAccessSlice(i, padded.length() - i);
         for (int j = 0; j < shorts.length; j++) {
-          assertEquals(shorts[j], whole.readShort(j * 2));
+          assertEquals(shorts[j], whole.readShort(j * 2L));
         }
         padded.close();
       }
